# services/swagger.yml
openapi: 3.0.0
servers:
  - url: http://127.0.0.1:8080/prod
info:
  version: 1.0.0
  title: Orbital Test Api
  description: Test api
components:
  schemas:
    Month:
      type: number
      minimum: 1
      maximum: 12
    Year:
      type: number
      minimum: 2000
      maximum: 3000
    MonthlyReport: 
      type: object
      required:
        - month
        - year
        - accountId
        - currencies
      properties:
        month:
          $ref: "#/components/schemas/Month"
        year:
          $ref: "#/components/schemas/Year"
        accountId:
          $ref: "#/components/schemas/UUID"
        currencies:
          type: array
          items:
            type: object
            required:
              - currency
              - transactions
            properties:
              currency:
                $ref: "#/components/schemas/SupportedCurrencies"
              transactions:
                type: array
                items:
                  $ref: "#/components/schemas/StatementTransaction"

    StatementTransaction:
      type: object
      required:
        - transactionId
        - type
        - amount
        - createdAt
      properties:
        transactionId:
          $ref: "#/components/schemas/UUID"
        type:
          $ref: "#/components/schemas/TransactionType"
        amount:
          $ref: "#/components/schemas/TransactionAmount"
        createdAt:
          $ref: "#/components/schemas/ISODate"

    SupportedCurrencies:
      type: string
      enum:
       - KWD
       - BHD
       - OMR
       - JOD
       - GIP
       - GBP
       - KYD
       - CHF
       - EUR
       - USD
    TransactionType:
      type: string
      enum:
       - INBOUND
       - OUTBOUND
    TransactionAmount:
      type: number
      minimum: 1
    BalanceAmount:
      type: number
    UUID:
      type: string
      pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
    ISODate:
      type: string
      pattern: ^^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}Z$
    
    CurrencyBalance:
      type: object
      required:
        - currency
        - amount
      properties:
        currency:
          $ref: "#/components/schemas/SupportedCurrencies"
        amount:
          $ref: "#/components/schemas/BalanceAmount"

    EventTypes:
      type: string
      enum:
       - account-created
       - transaction-created

    AccountPartial:
      type: object
      required:
        - supportedCurrencies
      properties:
        supportedCurrencies:
          type: array
          items:
            $ref: "#/components/schemas/SupportedCurrencies"

    Account:
      allOf:
        - $ref: "#/components/schemas/AccountPartial"
        - type: object
          required:
            - accountId
            - supportedCurrencies
          properties:
            accountId:
              $ref: "#/components/schemas/UUID"

    AccountWithBalance:
      allOf:
        - $ref: "#/components/schemas/Account"
        - type: object
          properties:
            balances:
              type: array
              items:
                $ref: "#/components/schemas/CurrencyBalance"
    TransactionPartial:
        type: object
        required:
          - currency
          - type
          - amount
        properties:
          currency:
            $ref: "#/components/schemas/SupportedCurrencies"
          type:
            $ref: "#/components/schemas/TransactionType"
          amount:
            $ref: "#/components/schemas/TransactionAmount"
    Transaction:
      allOf:
        - $ref: "#/components/schemas/TransactionPartial"
        - type: object
          required:
            - transactionId
            - createdAt
          properties:
            transactionId:
              $ref: "#/components/schemas/UUID"
            createdAt:
              $ref: "#/components/schemas/ISODate"

    CreateAccountRequest:
      type: object
      properties:   
        supportedCurrencies:
          type: array
          items:
            $ref: "#/components/schemas/SupportedCurrencies"

    CreateAccountTransactionRequest:
      $ref: "#/components/schemas/TransactionPartial"

    UpdateAccountRequest:
      type: object
      properties:   
        supportedCurrencies:
          type: array
          items:
            $ref: "#/components/schemas/SupportedCurrencies"

  requestBodies:
    CreateAccountRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateAccountRequest"

    CreateAccountTransactionRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateAccountTransactionRequest"
            
    UpdateAccountRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdateAccountRequest"
            
  responses:
    CreateAccountResponse:
      description: Account created with success
      content: 
        application/json:
          schema:
            $ref: "#/components/schemas/Account"

    GetAccountResponse:
      description: Account retrieved with success
      content: 
        application/json:
          schema:
            $ref: "#/components/schemas/AccountWithBalance"

    PatchAccountResponse:
      description: Account retrieved with success
      content: 
        application/json:
          schema:
            $ref: "#/components/schemas/Account"

    CreateAccountTransactionResponse:
      description: Account transaction created with success
      content: 
        application/json:
          schema:
            $ref: "#/components/schemas/Transaction"

    GetAccountTransactionsResponse:
      description: Account transactions retrieved with success
      content: 
        application/json:
          schema:
            type: array
            items: 
              $ref: "#/components/schemas/Transaction"

    GetMonthlyReportResponse:
      description: Monlty report for an account
      content: 
        application/json:
          schema:
            $ref: "#/components/schemas/MonthlyReport"
    
  parameters:
    AccountIdPathParameter:
      name: account_id
      required: true
      in: path
      description: The target account
      schema:
        $ref: "#/components/schemas/UUID"   
    MonthPathParameter:
      name: month
      required: true
      in: path
      schema:
        $ref: "#/components/schemas/Month"
    YearPathParameter:
      name: year
      required: true
      in: path
      schema:
        $ref: "#/components/schemas/Year"
      
paths:
  /v1/accounts:
    post:
      tags:
        - Account Service
      description: Create a new account  
      requestBody:
        $ref: "#/components/requestBodies/CreateAccountRequest"
      responses:
        '200':
          $ref: "#/components/responses/CreateAccountResponse"

  /v1/accounts/{account_id}:
    get:      
      tags:
        - Account Service
      description: Returns the details of an account   
      parameters:
          - $ref: "#/components/parameters/AccountIdPathParameter"        
      responses:
        '200':
          $ref: "#/components/responses/GetAccountResponse"
    patch:      
      tags:
        - Account Service
      description: Updates the details of an account  
      requestBody:
        $ref: "#/components/requestBodies/CreateAccountRequest"
      parameters:
          - $ref: "#/components/parameters/AccountIdPathParameter"        
      responses:
        '200':
          $ref: "#/components/responses/PatchAccountResponse"

  /v1/accounts/{account_id}/transactions:
    post:
      tags:
        - Account Service
      description: Create a new account transaction  
      requestBody:
        $ref: "#/components/requestBodies/CreateAccountTransactionRequest"
      parameters:
          - $ref: "#/components/parameters/AccountIdPathParameter"    
      responses:
        '200':
          $ref: "#/components/responses/CreateAccountTransactionResponse"
    get:      
      tags:
        - Account Service
      description: Returns a list of account transactions 
      parameters:
          - $ref: "#/components/parameters/AccountIdPathParameter"        
      responses:
        '200':
          $ref: "#/components/responses/GetAccountTransactionsResponse"

  /v1/reports/{account_id}/{year}/{month}:
    get:      
      tags:
        - Reports Service
      description: Returns a list of account transactions 
      parameters:
          - $ref: "#/components/parameters/AccountIdPathParameter"        
          - $ref: "#/components/parameters/YearPathParameter"        
          - $ref: "#/components/parameters/MonthPathParameter"        
      responses:
        '200':
          $ref: "#/components/responses/GetMonthlyReportResponse"
    
    
    